+++ pesummary/core/command_line.py
@@ -195,8 +195,8 @@ def command_line():
                         nargs='+', default=None)
     parser.add_argument("-v", "--verbose", action="store_true",
                         help="print useful information for debugging purposes")
-    parser.add_argument("--save_to_hdf5", action="store_true",
-                        help="save the meta file in hdf5 format", default=False)
+    parser.add_argument("--save_to_json", action="store_true",
+                        help="save the meta file in json format", default=False)
     parser.add_argument("--custom_plotting", dest="custom_plotting",
                         help=("python file containing functions for custom "
                               "plotting"), default=None)